requirements.txt
asgiref==3.8.1
Django==5.1.5
django-cors-headers==4.6.0
django-filter==24.3
djangorestframework==3.15.2
djangorestframework_simplejwt==5.4.0
msgpack==1.1.0
pillow==11.1.0
PyJWT==1.7.1
redis==5.2.1
sqlparse==0.5.3
